feat: WIP refactored the code to split out the note form from create form component
