+++ package.json
@@ -24,6 +24,7 @@
 		"framer-motion": "^6.5.1",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
+		"react-icons": "^4.7.1",
 		"react-markdown": "^8.0.4",
 		"react-router-dom": "^6.4.5",
 		"uuid": "^9.0.0",
